+++ projects/onlyoffice/document-editor-angular/src/lib/model/config.ts
@@ -73,6 +73,9 @@ export interface IConfig {
     };
     createUrl?: string;
     lang?: string;
+    /**
+     * @deprecated Deprecated since version 8.2, please use the region parameter instead.
+      */
     location?: string;
     mode?: string;
     recent?: any[];
@@ -102,6 +105,10 @@ export interface IConfig {
        * @deprecated Deprecated since version 7.1, please use the document.permissions.chat parameter instead.
        */
       chat?: boolean;
+      close: {
+        visible: boolean;
+        text: string;
+      }
       /**
        * @deprecated Deprecated since version 6.3, please use the document.permissions.editCommentAuthorOnly and document.permissions.deleteCommentAuthorOnly fields instead.
        */
@@ -132,8 +139,10 @@ export interface IConfig {
       logo?: {
         image?: string;
         imageDark?: string;
+        imageLight?: string;
         imageEmbedded?: string;
         url?: string;
+        visible?: boolean;
       };
       macros?: boolean;
       macrosMode?: string;
@@ -179,6 +188,10 @@ export interface IConfig {
     };
     plugins?: {
       autostart?: string[];
+      options?: {
+        all?: Object;
+        pluginGuid: Object;
+      };
       pluginsData?: string[];
       /**
        * @deprecated Deprecated since version 4.3, please use the absolute URLs in pluginsData field.
@@ -224,7 +237,9 @@ export interface IConfig {
     onRequestSelectSpreadsheet?: (event: object) => void;
     onRequestSendNotify?: (event: object) => void;
     onRequestSharingSettings?: (event: object) => void;
+    onRequestStartFilling: (event: object) => void;
     onRequestUsers?: (event: object) => void;
+    onSubmit?: (event: object) => void;
     onWarning?: (event: object) => void;
   };
 }